+++ GDX/Runtime/Developer/ObjectInfos/ShaderObjectInfo.cs
@@ -15,19 +15,7 @@ namespace GDX.Developer.ObjectInfos
         public int PassCount;
 #endif
 
-        public override void Populate(Object targetObject)
-        {
-            base.Populate(targetObject);
-            Shader shaderAsset = (Shader)targetObject;
-
-            // Useful shader information
-#if UNITY_2019_1_OR_NEWER
-            PassCount = shaderAsset.passCount;
-#endif
-            IsSupported = shaderAsset.isSupported;
-        }
-
-        public override void Populate(Object targetObject, TransientReference reference)
+        public override void Populate(Object targetObject, TransientReference reference = null)
         {
             base.Populate(targetObject, reference);
             Shader shaderAsset = (Shader)targetObject;