+++ GDX/Runtime/Developer/Report.cs
@@ -3,53 +3,117 @@
 // See the LICENSE file in the project root for more information.
 
 using System;
+using System.IO;
+using System.Text;
+using UnityEngine;
+using UnityEngine.Profiling;
+using UnityEngine.SceneManagement;
 
 namespace GDX.Developer
 {
-    /// <summary>
-    ///     A set of functionality used to generate reports.
-    /// </summary>
-    public static class Report
+    public class Report
     {
         /// <summary>
-        ///     The number of characters to be considered a line in a report.
+        ///     The name of the scene that was known to the <see cref="UnityEngine.SceneManagement" /> as being the active scene
+        ///     when this <see cref="ResourcesReport" /> was created.
         /// </summary>
-        public const int CharacterWidth = 80;
+        public readonly string ActiveScene;
 
         /// <summary>
-        ///     A <see cref="string" /> array used to represent the end of a line for splitting purposes.
+        ///     The time of creation for the <see cref="ResourcesReport" />.
         /// </summary>
-        public static readonly string[] NewLineSplit = {Environment.NewLine};
+        public readonly DateTime Created;
 
         /// <summary>
-        ///     Create a sized divider string for use in generating reports.
+        ///     The size of the Mono heap when the <see cref="ResourcesReport" /> was created.
         /// </summary>
-        /// <param name="divider">The optional character to use as the divider.</param>
-        /// <returns>A sized string to be used as a divider.</returns>
-        public static string CreateDivider(char divider = '-')
+        /// <remarks>This is cached so that the <see cref="ResourcesReport" /> does not effect this value.</remarks>
+        public readonly long MonoHeapSize;
+
+        /// <summary>
+        ///     The amount of the Mono heap used when the <see cref="ResourcesReport" /> was created.
+        /// </summary>
+        /// <remarks>This is cached so that the <see cref="ResourcesReport" /> does not effect this value.</remarks>
+        public readonly long MonoUsedSize;
+
+        /// <summary>
+        ///     The platform that the <see cref="ResourcesReport" /> was created on.
+        /// </summary>
+        public readonly RuntimePlatform Platform;
+
+        /// <summary>
+        ///     Unity's allocated native memory for the graphics driver (in bytes).
+        /// </summary>
+        public readonly long UnityGraphicsDriverAllocatedMemory;
+
+        /// <summary>
+        ///     Unity's total allocated memory (in bytes).
+        /// </summary>
+        public readonly long UnityTotalAllocatedMemory;
+
+        /// <summary>
+        ///     Unity's total reserved memory (in bytes).
+        /// </summary>
+        public readonly long UnityTotalReservedMemory;
+
+        /// <summary>
+        ///     Unity's total unused reserved memory (in bytes).
+        /// </summary>
+        public readonly long UnityTotalUnusedReservedMemory;
+
+        /// <summary>
+        ///     Unity's used portion of the heap (in bytes).
+        /// </summary>
+        public readonly long UnityUsedHeapSize;
+
+        public Report()
         {
-            return "".PadRight(CharacterWidth, divider);
+            ActiveScene = SceneManager.GetActiveScene().name;
+            Platform = Application.platform;
+            MonoUsedSize = Profiler.GetMonoUsedSizeLong();
+            MonoHeapSize = Profiler.GetMonoHeapSizeLong();
+            Created = DateTime.Now;
+
+            UnityUsedHeapSize = Profiler.usedHeapSizeLong;
+            UnityTotalAllocatedMemory = Profiler.GetTotalAllocatedMemoryLong();
+            UnityTotalReservedMemory = Profiler.GetTotalReservedMemoryLong();
+            UnityGraphicsDriverAllocatedMemory = Profiler.GetAllocatedMemoryForGraphicsDriver();
+            UnityTotalUnusedReservedMemory = Profiler.GetTotalUnusedReservedMemoryLong();
         }
 
         /// <summary>
-        ///     Create a header with <paramref name="title" /> with repeated <paramref name="decorator" />s on the sides, filling
-        ///     out to <see cref="CharacterWidth" />.
+        ///     Output the report format as an array of <see cref="string" />.
         /// </summary>
-        /// <param name="title">The text to be treated as the title for the header.</param>
-        /// <param name="decorator">The optional character to be used as the decorator.</param>
-        /// <returns>A sized string to be used as a header.</returns>
-        public static string CreateHeader(string title, char decorator = '=')
+        /// <remarks>It is usually best to provide a <see cref="StringBuilder" /> or <see cref="StreamWriter" /> instead.</remarks>
+        /// <returns>A generated report as an array of <see cref="string" />.</returns>
+        public string[] Output()
         {
-            string workingTitle = $" {title.Trim()} ";
+            StringBuilder builder = new StringBuilder();
+            return Output(ref builder)
+                ? builder.ToString().Split(ReportBuilder.NewLineSplit, StringSplitOptions.None)
+                : null;
+        }
 
-            int titleWidth = workingTitle.Length;
-            int decoratorSideWidth = (CharacterWidth - titleWidth) / 2;
+        /// <summary>
+        ///     Output the report format utilizing the provided <paramref name="writer" />, optionally limiting the
+        ///     write buffers by <paramref name="bufferSize" />.
+        /// </summary>
+        /// <param name="writer"></param>
+        /// <param name="bufferSize"></param>
+        /// <returns>true/false if the report was successfully written to the provided <paramref name="writer" />.</returns>
+        public bool Output(ref StreamWriter writer, int bufferSize = -1)
+        {
+            StringBuilder builder = new StringBuilder();
+            if (Output(ref builder))
+            {
+                //writer.Write();
+            }
+            return false;
+        }
 
-            // Pad left side first to ensure it is always the most accurate in length
-            workingTitle = workingTitle.PadLeft(titleWidth + decoratorSideWidth, decorator);
-            workingTitle = workingTitle.PadRight(CharacterWidth, decorator);
-
-            return workingTitle;
+        public virtual bool Output(ref StringBuilder builder)
+        {
+            return false;
         }
     }
 }